Fix: "Unable to find a medium containing a live file system #68"
On Ubuntu dash is the default shell linked onto /bin/sh. The dash shell
provides a quite different command syntax that does not support some of
the features that are required during the installation of the initrd.

As Ansible is always using the default shell for "shell" commands, if
not explicitely defined, it should be set on all commands that need
bash-specific syntax.